refactor: move StreamCreator to types/open-next
And make onFinish optional
