+++ .changeset/big-terms-boil.md
@@ -1,0 +1,5 @@
+---
+"@opennextjs/aws": minor
+---
+
+refactor: move StreamCreator to types/open-next
