+++ packages/open-next/src/adapters/image-optimization-adapter.ts
@@ -9,7 +9,6 @@ import path from "node:path";
 import type { Writable } from "node:stream";
 
 import { loadBuildId, loadConfig } from "config/util.js";
-import type { StreamCreator } from "http/openNextResponse.js";
 import { OpenNextNodeResponse } from "http/openNextResponse.js";
 // @ts-ignore
 import { defaultConfig } from "next/dist/server/config-shared";
@@ -19,7 +18,11 @@ import {
 } from "next/dist/server/image-optimizer";
 // @ts-ignore
 import type { NextUrlWithParsedQuery } from "next/dist/server/request-meta";
-import type { InternalEvent, InternalResult } from "types/open-next.js";
+import type {
+  InternalEvent,
+  InternalResult,
+  StreamCreator,
+} from "types/open-next.js";
 import { emptyReadableStream, toReadableStream } from "utils/stream.js";
 
 import { createGenericHandler } from "../core/createGenericHandler.js";
